requirements.txt
# Core Django
Django>=5.0,<6.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3

# Database
psycopg2-binary>=2.9

# API Documentation
drf-spectacular>=0.27

# Configuration & Environment
python-decouple>=3.8
django-cors-headers>=4.3

# Development Tools
django-extensions>=3.2